fix(github): use didgen instead of fmt

+++ plugins/github/models/run.go
@@ -27,7 +27,7 @@ type GithubRun struct {
 	common.NoPKModel
 	ConnectionId     uint64     `gorm:"primaryKey"`
 	RepoId           int        `gorm:"primaryKey"`
-	ID               int64      `json:"id" gorm:"primaryKey;autoIncrement:false"`
+	ID               int        `json:"id" gorm:"primaryKey;autoIncrement:false"`
 	Name             string     `json:"name" gorm:"type:varchar(255)"`
 	NodeID           string     `json:"node_id" gorm:"type:varchar(255)"`
 	HeadBranch       string     `json:"head_branch" gorm:"type:varchar(255)"`
